refactor: context 파일 등 파일명 및 변경된 위치 적용
