+++ src/advanced/constants/index.ts
@@ -1,16 +1,6 @@
-import { DOM_IDS } from "./domIds";
-import { ALERT_MESSAGES } from "./alertMessages";
-import { DISCOUNT_RATES } from "./discountRates";
-import { STOCKS } from "./stocks";
-import { ITEMS_REQUIRED_FOR_DISCOUNT } from "./itemsForDiscount";
-import { INITIAL_QUANTITY, TUESDAY } from "./general";
-
-export {
-  DOM_IDS,
-  ALERT_MESSAGES,
-  DISCOUNT_RATES,
-  STOCKS,
-  ITEMS_REQUIRED_FOR_DISCOUNT,
-  INITIAL_QUANTITY,
-  TUESDAY,
-};
+export * from "./domIds";
+export * from "./alertMessages";
+export * from "./discountRates";
+export * from "./stocks";
+export * from "./itemsForDiscounts";
+export * from "./general";
