+++ src/advanced/contexts/CartProvider.tsx
@@ -1,7 +1,9 @@
-import { createContext, useContext } from "react";
-import { useCart } from "./CartContext";
+import { createContext } from "react";
+import { useCart } from "advanced/hooks/useCart";
 
-const CartContext = createContext<ReturnType<typeof useCart> | null>(null);
+export const CartContext = createContext<ReturnType<typeof useCart> | null>(
+  null
+);
 
 export const CartProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
@@ -9,13 +11,4 @@ export const CartProvider: React.FC<{ children: React.ReactNode }> = ({
   const cart = useCart();
 
   return <CartContext.Provider value={cart}>{children}</CartContext.Provider>;
-};
-
-// Hook for accessing the CartContext
-export const useCartContext = () => {
-  const context = useContext(CartContext);
-  if (!context) {
-    throw new Error("useCartContext must be used within a CartProvider");
-  }
-  return context;
 };
